feat: tee_verifier_input_producer
Extract all data needed to re-execute and verify an L1Batch without accessing
the DB and/or the object store.

For testing purposes, the L1 batch is re-executed immediately for now.
Eventually, this component will only extract the inputs and send them to another
machine over a "to be defined" channel, e.g., save them to an object store.

Co-authored-by: Thomas Knauth <tk@matterlabs.dev>
Signed-off-by: Harald Hoyer <harald@matterlabs.dev>